+++ strawberry/channels/handlers/graphql_ws_handler.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from contextlib import suppress
-from typing import TYPE_CHECKING, Any, Optional
+from typing import TYPE_CHECKING, Any, Callable, Optional
 
 from strawberry.subscriptions import GRAPHQL_WS_PROTOCOL
 from strawberry.subscriptions.protocols.graphql_ws.handlers import BaseGraphQLWSHandler
@@ -19,8 +19,8 @@ class GraphQLWSHandler(BaseGraphQLWSHandler):
         debug: bool,
         keep_alive: bool,
         keep_alive_interval: float,
-        get_context,
-        get_root_value,
+        get_context: Callable,
+        get_root_value: Callable,
         ws: ChannelsWSConsumer,
     ):
         super().__init__(None, schema, debug, keep_alive, keep_alive_interval)
@@ -53,7 +53,7 @@ class GraphQLWSHandler(BaseGraphQLWSHandler):
     async def handle_request(self) -> Any:
         await self._ws.accept(subprotocol=GRAPHQL_WS_PROTOCOL)
 
-    async def handle_disconnect(self, code) -> None:
+    async def handle_disconnect(self, code: int) -> None:
         if self.keep_alive_task:
             self.keep_alive_task.cancel()
             with suppress(BaseException):