+++ tests/websockets/test_graphql_transport_ws.py
@@ -1,5 +1,6 @@
 import asyncio
 import json
+import sys
 import time
 from datetime import timedelta
 from typing import AsyncGenerator, Type
@@ -12,6 +13,7 @@ except ImportError:
 
 import pytest
 import pytest_asyncio
+from pytest_mock import MockerFixture
 
 from strawberry.subscriptions import GRAPHQL_TRANSPORT_WS_PROTOCOL
 from strawberry.subscriptions.protocols.graphql_transport_ws.types import (
@@ -154,6 +156,41 @@ async def test_connection_init_timeout_cancellation(
             payload={"data": {"debug": {"isConnectionInitTimeoutTaskDone": True}}},
         ).as_dict()
     )
+
+
+@pytest.mark.skipif(
+    sys.version_info < (3, 8),
+    reason="Task name was introduced in 3.8 and we need it for this test",
+)
+async def test_close_twice(
+    mocker: MockerFixture, request, http_client_class: Type[HttpClient]
+):
+    test_client = http_client_class()
+    test_client.create_app(connection_init_wait_timeout=timedelta(seconds=0.25))
+
+    async with test_client.ws_connect(
+        "/graphql", protocols=[GRAPHQL_TRANSPORT_WS_PROTOCOL]
+    ) as ws:
+        transport_close = mocker.patch.object(ws, "close")
+
+        await ws.send_json(ConnectionInitMessage(payload=None).as_dict())
+        # Yield control so that ._close can be called
+        await asyncio.sleep(0)
+
+        for t in asyncio.all_tasks():
+            if (
+                t.get_coro().__qualname__
+                == "BaseGraphQLTransportWSHandler.handle_connection_init_timeout"
+            ):
+                # The init timeout task should be cancelled
+                with pytest.raises(asyncio.CancelledError):
+                    await t
+
+        await ws.receive(timeout=0.5)
+        assert ws.closed
+        assert ws.close_code == 4400
+        ws.assert_reason("Invalid connection init payload")
+        transport_close.assert_not_called()
 
 
 async def test_too_many_initialisation_requests(ws: WebSocketClient):